rename org.eu.fedcampus to dev.flower.flower_tflite
