+++ examples/android-kotlin/client/flower_tflite/src/androidTest/java/dev/flower/flower_tflite/ExampleInstrumentedTest.kt
@@ -1,12 +1,10 @@
-package org.eu.fedcampus
+package dev.flower.flower_tflite
 
-import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
-
+import androidx.test.platform.app.InstrumentationRegistry
+import org.junit.Assert.*
 import org.junit.Test
 import org.junit.runner.RunWith
-
-import org.junit.Assert.*
 
 /**
  * Instrumented test, which will execute on an Android device.
@@ -19,6 +17,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("org.eu.fedcampus.test", appContext.packageName)
+        assertEquals("dev.flower.flower_tflite.test", appContext.packageName)
     }
 }
