+++ examples/android-kotlin/client/flower_tflite/src/main/java/dev/flower/flower_tflite/helpers/loss.kt
@@ -1,6 +1,6 @@
-package org.eu.fedcampus.train.helpers
+package dev.flower.flower_tflite.helpers
 
-import org.eu.fedcampus.train.Sample
+import dev.flower.flower_tflite.Sample
 import kotlin.math.ln
 
 fun <X> negativeLogLikelihoodLoss(
